+++ src/modules/account/views/ViewAccountDetails.vue
@@ -132,7 +132,7 @@ import qrcode from 'qrcode-generator';
 import PkPasswordModal from '@/modules/account/components/PkPasswordModal.vue'
 import PdfPasswordModal from '@/modules/account/components/PdfPasswordModal.vue'
 import DeleteAccountModal from '@/modules/account/components/DeleteAccountModal.vue'
-import { toSvg} from "jdenticon";
+import {toSvg} from "jdenticon";
 
 export default {
   name: "ViewAccountDetails",
@@ -154,8 +154,22 @@ export default {
     // get account details
     var acc = walletState.currentLoggedInWallet.accounts.find((add) => add.address == p.address);
     const other_acc = walletState.currentLoggedInWallet.others.find((add) => add.address == p.address);
-    const svgString = ref(toSvg(acc.address, 100))
- 
+
+    let jdenticonconfig = {
+      hues: [211],
+      lightness: {
+          color: [0.32, 0.80],
+          grayscale: [0.17, 0.82]
+      },
+      saturation: {
+          color: 1.00,
+          grayscale: 0.00
+      },
+      backColor: "#fff"
+    };
+
+    const svgString = ref(toSvg(acc.address, 100, jdenticonconfig));
+
     if(!acc){
       if(other_acc)
       {
