+++ src/modules/services/submodule/assets/views/ViewServicesAssetsLinkToNamespace.vue
@@ -1,139 +1,155 @@
 <template>
-  <div class="flex justify-between text-sm">
-    <div><span class="text-gray-300">{{$t('services.assets')}} ></span> <span class="text-blue-primary font-bold">{{$t('services.linktonamespace')}}</span></div>
-    <div>
-      <router-link :to="{ name: 'ViewServices' }" class="font-bold">{{$t('services.allservices')}}</router-link>
+ <div>
+  <div class="flex cursor-pointer mt-8 ml-8 lg:ml-0 lg:absolute">
+    <img src='@/assets/img/chevron_left.svg'>
+    <router-link :to="{name: 'ViewServicesAssets'}" class='text-blue-primary text-xs mt-0.5'>Back</router-link>
+  </div>
+  <div class='w-10/12 ml-auto mr-auto'>
+    <div class="border filter shadow-lg xl:grid xl:grid-cols-3 mt-8" >
+      <div class="xl:col-span-2 p-6 lg:p-12">
+        <div class="lg:flex lg:justify-between lg:items-center">
+          <div class='font-semibold mb-4 inline-block mt-1'>Link to Namespace</div>
+          <div class="flex items-center">
+            <div v-html="svgString" class="inline-block" />
+            <div class="ml-2">
+              <div class="text-blue-primary text-xxs font-bold uppercase mb-1">Asset created in</div>
+              <div class="font-bold text-black text-sm">{{ selectedAccName }}</div>
+            </div>
+          </div>
+        </div>
+
+        <div v-if="showNoBalance" class="rounded-md bg-red-200 w-full p-2 flex items-center justify-center">
+          <div class="rounded-full w-5 h-5 border border-red-500 inline-block relative mr-2"><font-awesome-icon icon="times" class="text-red-500 h-3 w-3 absolute" style="top: 3px; left:4px"></font-awesome-icon></div><div class="inline-block text-xs">{{$t('accounts.insufficientbalance')}}</div>
+        </div>
+        <div v-else-if="isNotCosigner" class="rounded-md bg-yellow-200 w-full p-2 flex items-center justify-center">
+          <div class="rounded-full w-5 h-5 bg-yellow-100 inline-block relative mr-2"><font-awesome-icon icon="exclamation" class="text-yellow-500 h-3 w-3 absolute" style="top: 5px; left:7px"></font-awesome-icon></div><div class="inline-block text-xs">{{$t('accounts.cosigwarning2')}}</div>
+        </div>
+        <div class="border border-blue-primary p-4 bg-blue-100 flex items-center rounded mt-5">
+          <img src="@/modules/services/submodule/assets/img/icon-asset.svg">
+          <div class="ml-1">
+            <div class="uppercase text-blue-primary font-semibold text-xxs">Asset ID</div>
+            <div class="text-black text-sm font-bold">{{ selectAsset }}</div>
+          </div>
+        </div>
+        <div class="border border-gray-200 p-4 rounded mt-5">
+          <div class="lg:grid lg:grid-cols-2">
+            <div class="my-3">
+              <div class="text-xxs text-blue-primary uppercase mb-1 font-bold">CURRENT SUPPLY<img src="@/assets/img/icon-info.svg" class="inline-block ml-2 relative" style="top: -1px;" v-tooltip.bottom="'<tiptext>Maximum supply is 900T.<br>Example: 900,000,000,000,000</tiptext>'"></div>
+              <div class="text-black font-bold text-sm">{{ assetSupply }}</div>
+            </div>
+            <div class="my-3">
+              <div class="text-xxs text-blue-primary uppercase mb-1 font-bold">Divisibility<img src="@/assets/img/icon-info.svg" class="inline-block ml-2 relative" style="top: -1px;" v-tooltip.bottom="'<tiptext>Determines up to what decimal place the asset can be divided.<br><br>Maximum divisibility is 6.<br>Example: 0.000000</tiptext>'"></div>
+              <div class="text-black font-bold text-sm">{{ assetDivisibility }}</div>
+            </div>
+            <div class="my-3">
+              <div class="text-xxs text-blue-primary uppercase mb-1 font-bold">Transferable<img src="@/assets/img/icon-info.svg" class="inline-block ml-2 relative" style="top: -1px;" v-tooltip.bottom="'<tiptext>If you tick \'Transferable\',<br>asset can be transferred.</tiptext>'"></div>
+              <div class="uppercase text-black font-bold text-sm">{{ assetTransferable?'Yes':'No' }}</div>
+            </div>
+            <div class="my-3">
+              <div class="text-xxs text-blue-primary uppercase mb-1 font-bold">Supply Mutable<img src="@/assets/img/icon-info.svg" class="inline-block ml-2 relative" style="top: -1px;" v-tooltip.bottom="'<tiptext>If you tick \'Supply Mutable\',<br>supply can be changed.</tiptext>'"></div>
+              <div class="uppercase text-black font-bold text-sm">{{ assetMutable?'Yes':'No' }}</div>
+            </div>
+          </div>
+        </div>
+        <div class="lg:grid lg:grid-cols-2 mt-5">
+          <SelectLinkType title="modification type" class="lg:mr-4" v-model="selectAction" :disabled="disabledSelectAction" />
+          <!-- <SupplyInputClean :disabled="showNoBalance||isNotCosigner" v-model="supply" :balance="balanceNumber" placeholder="Supply Quantity to Increase" type="text" icon="coins" :showError="showSupplyErr" :errorMessage="(!supply)? $t('scriptvalues.requiredfield'): $t('accounts.insufficientbalance')" :decimal="Number(assetDivisibility)" ckass="lg:ml-4" /> -->
+          <SelectInputNamespace :action="selectAction" v-model="selectNamespace" address="selectedAccAdd" />
+        </div>
+        <div class="error error_box" v-if="err!=''">{{ err }}</div>
+        <div class="mt-4">
+          <div v-if="getMultiSigCosigner.list.length > 0">
+            <div class="text-tsm text-left mt-3">{{$t('transfer.cosigner')}}:
+              <span class="font-bold" v-if="getMultiSigCosigner.list.length == 1">{{ getMultiSigCosigner.list[0].name }} ({{$t('services.balance')}}: {{ Helper.amountFormatterSimple(getMultiSigCosigner.list[0].balance, 0) }} XPX) <span v-if="getMultiSigCosigner.list[0].balance < lockFundTotalFee" class="error">- {{$t('accounts.insufficientbalance')}}</span></span>
+              <div v-if="cosignerBalanceInsufficient" class="error">- {{$t('accounts.insufficientbalance')}}</div>
+            </div>
+          </div>
+        </div>
+      </div>
+      <div class="bg-navy-primary py-6 px-12 xl:col-span-1">
+        <div class="font-semibold text-xxs text-blue-primary">ACCOUNT CURRENT BALANCE</div>
+        <div class="flex text-gray-200 mb-5">
+          <span v-html="splitCurrency(balance)"></span>
+          <img src="@/modules/account/img/proximax-logo.svg" class='ml-1 h-5 w-5 mt-0.5'>
+        </div>
+        <div class="flex justify-between border-gray-600 border-b items-center text-gray-200 text-xs py-3">
+          <div class="font-semibold">Transaction Fee</div>
+          <div v-html="splitCurrency(transactionFee)"></div>
+        </div>
+        <div class="flex justify-between border-gray-600 border-b items-center text-gray-200 text-xs py-3">
+          <div class="font-semibold">Rental Fee</div>
+          <div v-html="splitCurrency(rentalFeeCurrency)"></div>
+        </div>
+        <div class="flex justify-between border-gray-600 border-b items-center text-gray-200 text-xs py-3" v-if="isMultiSig(selectedAccAdd)">
+          <div class="font-semibold">{{$t('accounts.lockfund')}}</div>
+          <div v-html="splitCurrency(lockFundCurrency)"></div>
+        </div>
+        <div class="flex justify-between border-gray-600 border-b items-center text-gray-200 text-xs py-3" v-if="isMultiSig(selectedAccAdd)">
+          <div class="font-semibold">{{$t('accounts.unconfirmed')}}</div>
+          <div v-html="splitCurrency(lockFundTxFee)"></div>
+        </div>
+        <div class="flex justify-between border-gray-600 text-white text-xs py-5">
+          <div class="font-bold uppercase">Total</div>
+          <div v-html="splitCurrency(totalFeeFormatted)"></div>
+        </div>
+        <div class='text-xs text-white mt-5'>Enter your password to continue</div>
+        <div class='text-xs text-gray-400 mt-0.5 mb-1.5' >For security, this is required before proceeding to payment.</div>
+        <PasswordInput :placeholder="$t('signin.enterpassword')" errorMessage="Wallet password is required" :showError="showPasswdError" v-model="walletPassword" :disabled="disabledPassword" />
+        <button type="submit" class="mt-3 w-full blue-btn py-4 disabled:opacity-50 disabled:cursor-auto text-white" :disabled="disableCreate" @click="linkNamespace">Link to Namespace</button>
+        <div class="text-center">
+          <router-link :to="{name: 'ViewServicesAssets'}" class='content-center text-xs text-white border-b-2 border-white'>Cancel</router-link>
+        </div>
+      </div>
     </div>
   </div>
-  <div class='mt-2 py-3 gray-line text-center px-0 lg:px-10 xl:px-80'>
-    <div class="md:col-span-3">
-      <form @submit.prevent="linkNamespace">
-        <fieldset class="w-full">
-          <div class="mb-5">
-            <div v-if="showNoBalance" class="border-2 rounded-3xl border-red-700 w-full h-24 text-center p-4">
-              <div class="h-5 text-center">
-                <div class="rounded-full w-8 h-8 border border-gray-500 inline-block relative"><font-awesome-icon icon="times" class="text-gray-500 h-5 w-5 absolute" style="top: 5px; left:4px"></font-awesome-icon></div><br>
-                <div class="inline-block text-tsm">{{$t('accounts.insufficientbalance')}}</div>
-              </div>
-            </div>
-            <div v-if="isNotCosigner" class="border-2 rounded-3xl border-yellow-400 w-full h-24 text-center p-4">
-              <div class="h-5 text-center">
-                <div class="rounded-full w-8 h-8 border border-yellow-500 inline-block relative"><font-awesome-icon icon="exclamation" class="text-yellow-500 h-5 w-5 absolute" style="top: 5px; left:11px"></font-awesome-icon></div><br>
-                <div class="inline-block text-tsm">You are not a cosigner to this account</div>
-              </div>
-            </div>
-            <div class="error error_box" v-if="err!=''">{{ err }}</div>
-            <div v-if="moreThanOneAccount" class="text-left p-4">
-              <div class="mb-1 cursor-pointer z-20 border-b border-gray-200" @click="showMenu = !showMenu">
-                <div class="font-bold text-xs">{{ selectedAccName }} <span v-if="isMultiSigBool" class="text-xs font-normal ml-2 inline-block py-1 px-2 rounded bg-blue-200 text-gray-800">{{$t('accounts.multisig')}}</span></div>
-                <div class="text-gray-400 mt-1 text-sm ">{{ selectedAccAdd }}</div>
-              </div>
-              <transition name="slide">
-              <div v-if="showMenu" class="z-10">
-                <div :key="item.address" :i="index" v-for="(item, index) in accounts" class="p-2 cursor-pointer" :class="item.name==selectedAccName?'bg-blue-primary text-white font-bold':'text-gray-800 bg-gray-50 optionDiv'" @click="changeSelection(item)" :title="'Address is ' + item.address">
-                  <div>{{ item.name }} <span v-if="isMultiSig(item.address)" class="text-xs font-normal ml-2 inline-block py-1 px-2 rounded bg-blue-200 text-gray-800">{{$t('accounts.multisig')}}</span></div>
-                </div>
-              </div>
-              </transition>
-              <input type="hidden" v-model="currentSelectedName">
-            </div>
-            <div v-else class="text-left p-4">
-              <div class="mb-1 z-20 border-b border-gray-200">
-                <div class="font-bold text-xs">{{ selectedAccName }} <span v-if="isMultiSigBool" class="text-xs font-normal ml-2 inline-block py-1 px-2 rounded bg-blue-200 text-gray-800">Multisig</span></div>
-                <div class="text-gray-400 mt-1 text-sm ">{{ selectedAccAdd }}</div>
-              </div>
-            </div>
-            <div v-if="getMultiSigCosigner.list.length > 0">
-              <div class="text-tsm text-left ml-4">Cosigner:
-                <span class="font-bold" v-if="getMultiSigCosigner.list.length == 1">{{ getMultiSigCosigner.list[0].name }} (Balance: {{ getMultiSigCosigner.list[0].balance }} XPX) <span v-if="getMultiSigCosigner.list[0].balance < lockFundTotalFee" class="error">- Insufficient balance</span></span>
-                <span class="font-bold" v-else><select v-model="cosignerAddress"><option v-for="(cosigner, item) in getMultiSigCosigner.list" :value="cosigner.address" :key="item">{{ cosigner.name }} (Balance: {{ cosigner.balance }} XPX)</option></select></span>
-                <div v-if="cosignerBalanceInsufficient" class="error">- Insufficient balance</div>
-              </div>
-            </div>
-          </div>
-          <div class="text-left p-3 pb-0 border-l-8 border-gray-100 mb-5">
-            <div class="bg-gray-100 rounded-2xl p-3">
-              <div class="inline-block mr-4 text-tsm"><img src="@/assets/img/icon-prx-xpx-blue.svg" class="w-5 inline mr-1">{{$t('services.balance')}}: <span class="text-xs">{{ balance }} XPX</span></div>
-            </div>
-          </div>
-          <SelectInputPlugin showSelectTitleProp="true" placeholder="Select action" errorMessage="" v-model="selectAction" :disabled="disabledAction" :options="actions"  />
-          <SelectInputPlugin showSelectTitleProp="true" placeholder="Select namespace" errorMessage="" ref="selectNamespaceRef" :disabled="disabledNamespaceSelection" noOptionsText="No namespace for this account" v-model="selectNamespace" :options="namespaceOptions" :selectedAddress="selectedAccAdd" :selectedAction="selectAction" @show-selection="namespaceSelected" @clear-selection="clearNamespaceSelection" />
-          <SelectInputPlugin v-show="selectAction=='link'" showSelectTitleProp="true" placeholder="Select asset" errorMessage="" ref="selectAssetRef" :disabled="disabledAssetSelection" noOptionsText="No asset for this account" v-model="selectAsset" :options="assetOptions" @show-selection="assetSelected" />
-          <div class="rounded-2xl bg-gray-100 p-5 mb-5">
-            <div class="inline-block mr-4 text-xs"><img src="@/assets/img/icon-prx-xpx-blue.svg" class="w-5 inline mr-1 text-gray-500">{{$t('namespace.transactionfee')}}: <span>{{ transactionFee }}</span> XPX</div>
-          </div>
-          <div class="p-4 rounded-xl bg-gray-100 mt-2 items-center w-full text-xs text-gray-800 mb-5" v-if="isMultiSig(selectedAccAdd)">
-            <div class="text-center">
-              <div class="inline-block">
-                <div class="flex">
-                  <img src="@/assets/img/icon-prx-xpx-blue.svg" class="w-5 inline-block mr-1 self-center">
-                  <div class="inline-block self-center text-left">
-                    <div>LockFund: {{ lockFundCurrency }} {{ currencyName }}</div>
-                    <div>Unconfirmed/Recommended Fee: {{ lockFundTxFee }} {{ currencyName }}</div>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-          <PasswordInput placeholder="Enter Wallet Password" :errorMessage="'Please enter wallet password'" :showError="showPasswdError" v-model="walletPassword" icon="lock" :disabled="disabledPassword" />
-          <div class="mt-10">
-            <button type="button" class="default-btn mr-5 focus:outline-none disabled:opacity-50" :disabled="disabledClear" @click="clearInput()">{{$t('signin.clear')}}</button>
-            <button type="button" class="default-btn py-1 disabled:opacity-50" :disabled="disableCreate" @click="linkNamespace">{{$t('welcome.create')}}</button>
-          </div>
-        </fieldset>
-      </form>
-    </div>
-  </div>
+</div>
 </template>
 <script>
 import { computed, ref, watch } from 'vue';
-// import { useRouter } from "vue-router";
+import { useRouter } from "vue-router";
 import PasswordInput from '@/components/PasswordInput.vue';
-import SelectInputPlugin from '@/components/SelectInputPlugin.vue';
-// import SelectInputNamespaceAsyncOptionPlugin from '@/modules/services/submodule/assets/components/SelectInputNamespaceAsyncOptionPlugin.vue';
+import SelectLinkType from '@/modules/services/submodule/assets/components/SelectLinkType.vue';
+import SelectInputNamespace from '@/modules/services/submodule/assets/components/SelectInputNamespace.vue';
+import { ChainProfileConfig } from "@/models/stores/";
+import { Wallet } from "@/models/wallet";
 import { walletState } from "@/state/walletState";
 import { networkState } from "@/state/networkState";
+import { Currency } from "@/models/currency";
 import { Helper } from '@/util/typeHelper';
+import { ChainUtils } from '@/util/chainUtils';
 import { AssetsUtils } from '@/util/assetsUtils';
-
+import { WalletUtils } from '@/util/walletUtils';
+import { toSvg } from "jdenticon";
+import { useI18n } from 'vue-i18n';
+import { useToast } from "primevue/usetoast";
+import Tooltip from 'primevue/tooltip';
 
 export default {
-  name: 'ViewMosaicLinkToNamespace',
+  name: 'ViewServicesAssetsLinkToNamespace',
+  directives: { 'tooltip': Tooltip },
   components: {
     PasswordInput,
-    SelectInputPlugin,
-    // SelectInputNamespaceAsyncOptionPlugin,
+    SelectLinkType,
+    SelectInputNamespace,
   },
-  setup(){
-    const selectNamespaceRef = ref(null);
-    const selectAssetRef = ref(null);
+  props: {
+    assetId: String,
+    address: String,
+  },
+  setup(props){
+    const {t} = useI18n();
+    const router = useRouter();
+    const toast = useToast();
+    let maxAmount = 9999999999.999999;
+
+    const currentNativeTokenName = computed(()=> networkState.currentNetworkProfile.network.currency.name);
+
+    const showSupplyErr = ref(false);
     const walletPassword = ref('');
     const err = ref('');
-    const showMenu = ref(false);
-    const currentSelectedName = ref('');
     const disabledPassword = ref(false);
-    const disabledClear = ref(false);
-
-    const disabledAction = ref(false);
-    const disabledAssetSelection = ref(true);
-    const disabledNamespaceSelection = ref(false);
-
-    const selectAction = ref('');
-    const actions = computed(() => {
-      let action = [];
-      action.push({value: 'link', label: 'Link'});
-      action.push({value: 'unlink', label: 'Unlink'});
-      return action;
-    });
-    const selectNamespace = ref('');
-
-    // const namespaceOption = computed(() => {
-    //   // console.log(selectedAccAdd.value);
-    //   const namespacesList = NamespacesUtils.listNamespacesToLink(selectedAccAdd.value, selectAction.value);
-    //   // console.log(namespacesList)
-    //   return namespacesList;
-    // });
-
+    const disabledSupply = ref(false);
+    const disabledSelectAction = ref(false);
     const passwdPattern = "^[^ ]{8,}$";
     const showPasswdError = ref(false);
 
@@ -141,16 +157,21 @@ export default {
     const cosignerAddress = ref('');
 
     const currencyName = computed(() => networkState.currentNetworkProfile.network.currency.name);
-
     const lockFund = computed(()=> Helper.convertToExact(networkState.currentNetworkProfileConfig.lockedFundsPerAggregate, networkState.currentNetworkProfile.network.currency.divisibility))
     const lockFundCurrency = computed(()=> Helper.convertToCurrency(networkState.currentNetworkProfileConfig.lockedFundsPerAggregate, networkState.currentNetworkProfile.network.currency.divisibility))
 
     const lockFundTxFee = ref(0.0445);
     const lockFundTotalFee = computed(()=> lockFund.value + lockFundTxFee.value);
 
+
     const disableCreate = computed(() => !(
-      walletPassword.value.match(passwdPattern)
+      walletPassword.value.match(passwdPattern) && (supply.value > 0)
     ));
+
+    const selectedAccName = ref(walletState.currentLoggedInWallet.selectDefaultAccount().name);
+    const selectedAccAdd = ref(walletState.currentLoggedInWallet.selectDefaultAccount().address);
+    const balance = ref(Helper.toCurrencyFormat(walletState.currentLoggedInWallet.selectDefaultAccount().balance, networkState.currentNetworkProfile.network.currency.divisibility));
+    const balanceNumber = ref(maxAmount);
 
     const isMultiSig = (address) => {
       const account = walletState.currentLoggedInWallet.accounts.find((account) => account.address == address);
@@ -164,78 +185,86 @@ export default {
       return isMulti;
     };
 
-    const selectedAccName = ref(walletState.currentLoggedInWallet.selectDefaultAccount().name);
-    const selectedAccAdd = ref(walletState.currentLoggedInWallet.selectDefaultAccount().address);
-    const balance = ref(Helper.toCurrencyFormat(walletState.currentLoggedInWallet.selectDefaultAccount().balance, networkState.currentNetworkProfile.network.currency.divisibility));
-    const isMultiSigBool = ref(isMultiSig(walletState.currentLoggedInWallet.selectDefaultAccount().address));
+    const isMultiSigBool = ref(isMultiSig(selectedAccAdd.value));
 
     const showNoBalance = ref(false);
-    const isNotCosigner = computed(() => getMultiSigCosigner.value.list.length == 0 && isMultiSig(selectedAccAdd.value));
+    const showNoAsset = ref(false);
+    const isNotCosigner = computed(() => getMultiSigCosigner.value.list.length == 0 && isMultiSig(selectedAccAdd.value) && !showNoAsset.value);
 
-    const accounts = computed( () => {
-      if(walletState.currentLoggedInWallet){
-        if(walletState.currentLoggedInWallet.others){
-          const concatOther = walletState.currentLoggedInWallet.accounts.concat(walletState.currentLoggedInWallet.others)
-          return concatOther;
-        } else{
-          return walletState.currentLoggedInWallet.accounts;
-        }
-      } else{
-        return [];
+    const supply = ref('0');
+
+    // get asset id
+    let other;
+    let account = walletState.currentLoggedInWallet.accounts.find((account) => Helper.createAddress(account.address).pretty() == props.address);
+    if(account != undefined){
+      selectedAccName.value = account.name;
+      selectedAccAdd.value = Helper.createAddress(account.address).pretty();
+      balance.value = account.balance;
+    }else{
+      other = walletState.currentLoggedInWallet.others.find((account) => account.address == props.address);
+      if(other != undefined){
+        selectedAccName.value = other.name;
+        selectedAccAdd.value = Helper.createAddress(other.address).pretty();
+        balance.value = other.balance;
+      }else{
+        toast.add({severity:'error', detail: 'Addres is invalid', group: 'br', life: 3000});
+        router.push({ name: "ViewServicesAssets" });
       }
-    });
+    }
 
-    const moreThanOneAccount = computed(()=> (accounts.value.length > 1)?true:false);
+    let jdenticonconfig = {
+      hues: [211],
+      lightness: {
+          color: [0.32, 0.80],
+          grayscale: [0.17, 0.82]
+      },
+      saturation: {
+          color: 1.00,
+          grayscale: 0.00
+      },
+      backColor: "#fff"
+    };
+
+    const svgString = ref(toSvg(selectedAccAdd.value, 40, jdenticonconfig));
+
+    const selectAsset = ref('');
+    const assetDivisibility = ref(0);
+    const assetSupply = ref(0);
+    const assetTransferable = ref(false);
+    const assetMutable = ref(false);
+    const selectAction = ref('link');
+    const selectNamespace = ref('');
+
+    if(account || other){
+      let asset = account.assets.find( asset => asset.idHex === props.assetId);
+      if(asset != undefined){
+        selectAsset.value = asset.idHex;
+        assetTransferable.value = asset.transferable;
+        assetMutable.value = asset.supplyMutable;
+        assetDivisibility.value = asset.divisibility;
+        assetSupply.value = Helper.convertToCurrency(asset.supply, asset.divisibility);
+      }else{
+        toast.add({severity:'error', detail: 'Asset ID is invalid', group: 'br', life: 3000});
+        router.push({ name: "ViewServicesAssets" });
+      }
+    }
+
+    
 
     const transactionFee = ref('0.000000');
     const transactionFeeExact = ref(0);
+    const rentalFeeCurrency = computed(()=> Helper.convertToCurrency(networkState.currentNetworkProfileConfig.mosaicRentalFee, networkState.currentNetworkProfile.network.currency.divisibility) );
+
+    const ownerPublicAccount = ref(WalletUtils.createPublicAccount(walletState.currentLoggedInWallet.selectDefaultAccount().publicKey, networkState.currentNetworkProfile.network.type));
 
     const getMultiSigCosigner = computed(() => {
       return AssetsUtils.getCosignerList(selectedAccAdd.value);
     });
 
-    const changeSelection = (i) => {
-      selectNamespaceRef.value.clear();
-      selectAssetRef.value.clear();
-      selectedAccName.value = i.name;
-      selectedAccAdd.value = i.address;
-      balance.value = i.balance;
-      isMultiSigBool.value = isMultiSig(i.address);
-      if(isMultiSigBool.value){
-        if((balance.value < (transactionFeeExact.value + lockFundTotalFee.value)) && !isNotCosigner.value && !showNoAsset.value){
-          showNoBalance.value = true;
-        }else{
-          showNoBalance.value = false;
-        }
-      }else{
-        if((balance.value < transactionFeeExact.value) && !isNotCosigner.value && !showNoAsset.value){
-          showNoBalance.value = true;
-        }else{
-          showNoBalance.value = false;
-        }
-      }
-      showMenu.value = !showMenu.value;
-      currentSelectedName.value = i.name;
-    }
-
-    const assetOptions = computed(() => {
-      return AssetsUtils.getOwnedAssets(selectedAccAdd.value);
-    });
-
-    const namespaceOptions = computed(() => {
-      return AssetsUtils.listActiveNamespacesToLink(selectedAccAdd.value, selectAction.value);
-    });
-
-    const selectAsset = ref('');
-
-    const clearInput = () => {
-      walletPassword.value = '';
-      selectNamespaceRef.value.clear();
-      selectAssetRef.value.clear()
-    };
+    transactionFee.value = Helper.amountFormatterSimple(AssetsUtils.getMosaicSupplyChangeTransactionFee(networkState.currentNetworkProfile.network.type, networkState.currentNetworkProfile.generationHash, selectAsset.value, selectAction.value, supply.value, assetDivisibility.value), networkState.currentNetworkProfile.network.currency.divisibility);
+    transactionFeeExact.value = Helper.convertToExact(AssetsUtils.getMosaicSupplyChangeTransactionFee(networkState.currentNetworkProfile.network.type, networkState.currentNetworkProfile.generationHash, selectAsset.value, selectAction.value, supply.value, assetDivisibility.value), networkState.currentNetworkProfile.network.currency.divisibility);
 
     const linkNamespace = () => {
-      // console.log('Link namespace method here');
       let assetId;
       if(selectAction.value=='link'){
         assetId = selectAsset.value;
@@ -249,29 +278,25 @@ export default {
       }else{
         AssetsUtils.linkedNamespaceToAsset(selectedAccAdd.value, walletPassword.value, networkState.currentNetworkProfile.network.type, networkState.currentNetworkProfile.generationHash, assetId, selectNamespace.value, selectAction.value );
       }
-      clearInput();
     };
 
-    const clearNamespaceSelection = () => {
-      selectAssetRef.value.clear()
-      disabledAssetSelection.value = true;
-    };
+    watch(selectAction, (n) => {
+      if(selectAsset.value){
+        transactionFee.value = Helper.amountFormatterSimple(AssetsUtils.getMosaicSupplyChangeTransactionFee(networkState.currentNetworkProfile.network.type, networkState.currentNetworkProfile.generationHash, selectAsset.value, n, supply.value, assetDivisibility.value), networkState.currentNetworkProfile.network.currency.divisibility);
+        transactionFeeExact.value = Helper.convertToExact(AssetsUtils.getMosaicSupplyChangeTransactionFee(networkState.currentNetworkProfile.network.type, networkState.currentNetworkProfile.generationHash, selectAsset.value, n, supply.value, assetDivisibility.value), networkState.currentNetworkProfile.network.currency.divisibility);
+        balanceNumber.value = (n=='increase'?maxAmount:parseFloat(assetSupply.value));
+      }else{
+        balanceNumber.value = (n=='increase'?maxAmount:0);
+      }
+      showSupplyErr.value = supply.value>balanceNumber.value;
+    });
 
-    const namespaceSelected = () => {
-      disabledAssetSelection.value = false;
-    };
-
-    const assetSelected = () => {
-      transactionFee.value = Helper.amountFormatterSimple(AssetsUtils.getLinkAssetToNamespaceTransactionFee(networkState.currentNetworkProfile.network.type, networkState.currentNetworkProfile.generationHash, selectAsset.value, selectNamespace.value, selectAction.value ), networkState.currentNetworkProfile.network.currency.divisibility);
-      transactionFeeExact.value = Helper.convertToExact(AssetsUtils.getLinkAssetToNamespaceTransactionFee(networkState.currentNetworkProfile.network.type, networkState.currentNetworkProfile.generationHash, selectAsset.value, selectNamespace.value, selectAction.value), networkState.currentNetworkProfile.network.currency.divisibility);
-    };
-
-    const setFormInput = (isValidate) => {
-      disabledAction.value = isValidate;
-      disabledPassword.value = isValidate;
-      disabledNamespaceSelection.value = isValidate;
-      disabledAssetSelection.value = isValidate;
-    };
+    watch(supply, (n) => {
+      if(selectAsset.value){
+        transactionFee.value = Helper.amountFormatterSimple(AssetsUtils.getMosaicSupplyChangeTransactionFee(networkState.currentNetworkProfile.network.type, networkState.currentNetworkProfile.generationHash, selectAsset.value, selectAction.value, n, assetDivisibility.value), networkState.currentNetworkProfile.network.currency.divisibility);
+        transactionFeeExact.value = Helper.convertToExact(AssetsUtils.getMosaicSupplyChangeTransactionFee(networkState.currentNetworkProfile.network.type, networkState.currentNetworkProfile.generationHash, selectAsset.value, selectAction.value, n, assetDivisibility.value), networkState.currentNetworkProfile.network.currency.divisibility);
+      }
+    });
 
     const totalFee = computed(() => {
       // if multisig
@@ -281,6 +306,16 @@ export default {
         return transactionFeeExact.value;
       }
     });
+
+    const totalFeeFormatted = computed(() => {
+      return Helper.amountFormatterSimple(totalFee.value, 0);
+    });
+
+    const setFormInput = (isValidate) => {
+      disabledPassword.value = isValidate;
+      disabledSupply.value = isValidate;
+      disabledSelectAction.value = isValidate;
+    };
 
     watch(totalFee, (n) => {
       if(balance.value < n){
@@ -296,8 +331,12 @@ export default {
       }
     });
 
-    watch(selectAction, () => {
-      selectNamespaceRef.value.clear();
+    watch(showNoAsset, (n) => {
+      if(n){
+        setFormInput(true);
+      }else{
+        setFormInput(false);
+      }
     });
 
     watch(showNoBalance, (n) => {
@@ -331,50 +370,56 @@ export default {
       }
     });
 
-    return {
-      accounts,
-      selectNamespaceRef,
-      selectAssetRef,
-      moreThanOneAccount,
-      showMenu,
-      currentSelectedName,
+    const splitCurrency = (amount) => {
+      let split = amount.toString().split(".")
+      if (split[1]!=undefined){
+        return '<span class="font-semibold text-sm">' + split[0] + '</span>.<span class="font-semibold text-xs">' + split[1] + ' ' + currentNativeTokenName.value + '</span>';
+      }else{
+        return '<span class="font-semibold text-sm">' + split[0] + '</span> <span class="font-semibold text-xs">' + currentNativeTokenName.value + '</span>';
+      }
+    };
+
+    return{
       selectedAccName,
       selectedAccAdd,
       balance,
+      balanceNumber,
       showNoBalance,
-      lockFundCurrency,
       lockFundTxFee,
+      lockFundCurrency,
       lockFundTotalFee,
+      totalFeeFormatted,
+      showNoAsset,
+      showSupplyErr,
       err,
       walletPassword,
       disableCreate,
-      clearInput,
       showPasswdError,
-      changeSelection,
-      disabledAction,
-      disabledNamespaceSelection,
-      disabledAssetSelection,
+      supply,
       disabledPassword,
-      disabledClear,
+      disabledSupply,
       currencyName,
       isMultiSig,
       isMultiSigBool,
-      assetOptions,
-      namespaceOptions,
       selectAsset,
-      actions,
       selectAction,
-      selectNamespace,
       linkNamespace,
+      transactionFee,
+      rentalFeeCurrency,
+      transactionFeeExact,
+      assetSupply,
+      assetDivisibility,
+      assetTransferable,
+      assetMutable,
+      selectNamespace,
       getMultiSigCosigner,
       cosignerBalanceInsufficient,
       cosignerAddress,
       isNotCosigner,
-      clearNamespaceSelection,
-      namespaceSelected,
-      assetSelected,
-      transactionFee,
-      transactionFeeExact,
+      disabledSelectAction,
+      splitCurrency,
+      Helper,
+      svgString,
     }
   },
 
@@ -382,52 +427,4 @@ export default {
 </script>
 <style scoped lang="scss">
 
-.slide-enter-active {
-   -moz-transition-duration: 1s;
-   -webkit-transition-duration: 1s;
-   -o-transition-duration: 1s;
-   transition-duration: 1s;
-   -moz-transition-timing-function: ease-in;
-   -webkit-transition-timing-function: ease-in;
-   -o-transition-timing-function: ease-in;
-   transition-timing-function: ease-in;
-}
-
-.slide-leave-active {
-   -moz-transition-duration: 1s;
-   -webkit-transition-duration: 1s;
-   -o-transition-duration: 1s;
-   transition-duration: 1s;
-   -moz-transition-timing-function: cubic-bezier(0, 1, 0.5, 1);
-   -webkit-transition-timing-function: cubic-bezier(0, 1, 0.5, 1);
-   -o-transition-timing-function: cubic-bezier(0, 1, 0.5, 1);
-   transition-timing-function: cubic-bezier(0, 1, 0.5, 1);
-}
-
-.slide-enter-to, .slide-leave-from {
-  max-height: 1000px;
-  overflow: hidden;
-}
-
-.slide-enter-from, .slide-leave-to {
-   overflow: hidden;
-   max-height: 0;
-}
-
-.optionDiv:hover{
-  background: #D9EBFF;
-}
-
-.property-table{
-  > div {
-    margin-bottom: 7px;
-    > div{
-      display: inline-block;
-    }
-    > div:first-child{
-      width: 120px;
-      margin-right: 15px;
-    }
-  }
-}
 </style>
