--- src/modules/services/submodule/assets/views/ViewServicesAssetsModifySupplyChange_.vue
@@ -1,518 +1,0 @@
-<template>
-<div>
-  <div class="flex justify-between text-sm">
-    <div><span class="text-gray-300">{{$t('services.assets')}} ></span> <span class="text-blue-primary font-bold">{{$t('services.modifysupply')}}</span></div>
-    <div>
-      <router-link :to="{ name: 'ViewServices' }" class="font-bold">{{$t('services.allservices')}}</router-link>
-    </div>
-  </div>
-  <div class='mt-2 py-3 gray-line text-center md:grid md:grid-cols-8 lg:grid-cols-6'>
-    <div class="md:col-span-5 lg:col-span-4">
-      <form @submit.prevent="modifyMosaic">
-        <fieldset class="w-full">
-          <div class="mb-5">
-            <div v-if="showNoBalance" class="border-2 rounded-3xl border-red-700 w-full h-24 text-center p-4">
-              <div class="h-5 text-center">
-                <div class="rounded-full w-8 h-8 border border-gray-500 inline-block relative"><font-awesome-icon icon="times" class="text-gray-500 h-5 w-5 absolute" style="top: 5px; left:8px"></font-awesome-icon></div><br>
-                <div class="inline-block text-tsm">{{$t('accounts.insufficientbalance')}}</div>
-              </div>
-            </div>
-            <div v-if="showNoAsset" class="border-2 rounded-3xl border-red-700 w-full h-24 text-center p-4">
-              <div class="h-5 text-center">
-                <div class="rounded-full w-8 h-8 border border-gray-500 inline-block relative"><font-awesome-icon icon="times" class="text-gray-500 h-5 w-5 absolute" style="top: 5px; left:9px"></font-awesome-icon></div><br>
-                <div class="inline-block text-tsm">{{$t('services.assetmessage')}}</div>
-              </div>
-            </div>
-            <div v-if="isNotCosigner" class="border-2 rounded-3xl border-yellow-400 w-full h-24 text-center p-4">
-              <div class="h-5 text-center">
-                <div class="rounded-full w-8 h-8 border border-yellow-500 inline-block relative"><font-awesome-icon icon="exclamation" class="text-yellow-500 h-5 w-5 absolute" style="top: 5px; left:11px"></font-awesome-icon></div><br>
-                <div class="inline-block text-tsm">{{$t('accounts.cosigwarning2')}}</div>
-              </div>
-            </div>
-            <div class="error error_box" v-if="err!=''">{{ err }}</div>
-            <div v-if="moreThanOneAccount" class="text-left p-4">
-              <div class="mb-1 cursor-pointer z-20 border-b border-gray-200" @click="showMenu = !showMenu">
-                <div class="font-bold text-xs">{{ selectedAccName }} <span v-if="isMultiSigBool" class="text-xs font-normal ml-2 inline-block py-1 px-2 rounded bg-blue-200 text-gray-800">{{$t('accounts.multisig')}}</span></div>
-                <div class="text-gray-400 mt-1 text-sm ">{{ selectedAccAdd }}</div>
-              </div>
-              <transition name="slide">
-              <div v-if="showMenu" class="z-10">
-                <div :key="item.address" :i="index" v-for="(item, index) in accounts" class="p-2 cursor-pointer" :class="item.name==selectedAccName?'bg-blue-primary text-white font-bold':'text-gray-800 bg-gray-50 optionDiv'" @click="changeSelection(item)" :title="'Address is ' + item.address">
-                  <div>{{ item.name }} <span v-if="isMultiSig(item.address)" class="text-xs font-normal ml-2 inline-block py-1 px-2 rounded bg-blue-200 text-gray-800">{{$t('accounts.multisig')}}</span></div>
-                </div>
-              </div>
-              </transition>
-              <input type="hidden" v-model="currentSelectedName">
-            </div>
-            <div v-else class="text-left p-4">
-              <div class="mb-1 z-20 border-b border-gray-200">
-                <div class="font-bold text-xs">{{ selectedAccName }} <span v-if="isMultiSigBool" class="text-xs font-normal ml-2 inline-block py-1 px-2 rounded bg-blue-200 text-gray-800">{{$t('accounts.multisig')}}</span></div>
-                <div class="text-gray-400 mt-1 text-sm ">{{ selectedAccAdd }}</div>
-              </div>
-            </div>
-            <div v-if="getMultiSigCosigner.list.length > 0">
-              <div class="text-tsm text-left ml-4">{{$t('transfer.cosigner')}}:
-                <span class="font-bold" v-if="getMultiSigCosigner.list.length == 1">{{ getMultiSigCosigner.list[0].name }} ({{$t('services.balance')}}: {{ getMultiSigCosigner.list[0].balance }} XPX) <span v-if="getMultiSigCosigner.list[0].balance < lockFundTotalFee" class="error">- {{$t('accounts.insufficientbalance')}}</span></span>
-                <span class="font-bold" v-else><select v-model="cosignerAddress"><option v-for="(cosigner, item) in getMultiSigCosigner.list" :value="cosigner.address" :key="item">{{ cosigner.name }} ({{$t('services.balance')}}: {{ cosigner.balance }} XPX)</option></select></span>
-                <div v-if="cosignerBalanceInsufficient" class="error">- {{$t('accounts.insufficientbalance')}}</div>
-              </div>
-            </div>
-          </div>
-          <div class="text-left p-3 pb-0 border-l-8 border-gray-100 mb-5">
-            <div class="bg-gray-100 rounded-2xl p-3">
-              <div class="inline-block mr-4 text-tsm"><img src="@/assets/img/icon-prx-xpx-blue.svg" class="w-5 inline mr-1">{{$t('services.balance')}}: <span class="text-xs">{{ balance }} XPX</span></div>
-            </div>
-          </div>
-          <SelectInputPlugin showSelectTitleProp="true" :placeholder="$t('services.selectasset')" errorMessage="" ref="assetOption" :noOptionsText="$t('services.noasset')" v-model="selectAsset" :options="assetOptions" @show-selection="changeAsset" :disabled="disabledSelectAsset" @clear-selection="clearAsset" />
-          <SelectInputPlugin :selectDefault="selectIncreaseDecrease" showSelectTitleProp="true" :placeholder="$t('services.addminus')" errorMessage="" v-model="selectIncreaseDecrease" :options="increaseDecreaseOption()" :disabled="disabledSelectIncreaseDecrease" />
-          <SupplyInput :disabled="disabledSupply" v-model="supply" :title="$t('services.quantityoption')" :balance="balanceNumber" placeholder="Supply" type="text" icon="coins" :showError="showSupplyErr" :errorMessage="(!supply)?'Required Field':(balanceNumber?'Max. amount to decrease is '+ balanceNumber:'Select asset')" :decimal="assetDivisibility" />
-          <div class="rounded-2xl bg-gray-100 p-5 mb-5">
-            <div class="inline-block mr-4 text-xs"><img src="@/assets/img/icon-prx-xpx-blue.svg" class="w-5 inline mr-1 text-gray-500">{{$t('namespace.transactionfee')}} {{ transactionFee }} XPX</div>
-          </div>
-          <div class="p-4 rounded-xl bg-gray-100 mt-2 items-center w-full text-xs text-gray-800 mb-5" v-if="isMultiSig(selectedAccAdd)">
-            <div class="text-center">
-              <div class="inline-block">
-                <div class="flex">
-                  <img src="@/assets/img/icon-prx-xpx-blue.svg" class="w-5 inline-block mr-1 self-center">
-                  <div class="inline-block self-center text-left">
-                    <div>LockFund: {{ lockFundCurrency }} {{ currencyName }}</div>
-                    <div>Unconfirmed/Recommended Fee: {{ lockFundTxFee }} {{ currencyName }}</div>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-          <PasswordInput :placeholder="$t('accounts.inputpassword')" :errorMessage="$t('scriptvalues.enterpassword',{name: walletName })" :showError="showPasswdError" v-model="walletPassword" icon="lock" :disabled="disabledPassword" />
-          <div class="mt-10">
-            <button type="button" class="default-btn mr-5 focus:outline-none disabled:opacity-50" :disabled="disabledClear" @click="clearInput">{{$t('signin.clear')}}</button>
-            <button type="button" class="default-btn py-1 disabled:opacity-50" :disabled="disableCreate" @click="modifyMosaic">{{$t('welcome.create')}}</button>
-          </div>
-        </fieldset>
-      </form>
-    </div>
-    <div class="md:col-span-3 lg:col-span-2 pl-10 text-left text-tsm mt-5 md:mt-0 property-table">
-      <div class="mb-2 border-b border-gray-300 pb-2 italic">{{$t('services.properties')}}</div>
-      <div>
-        <div class="italic text-right text-xs">{{$t('services.supply')}}:</div>
-        <div>{{ assetSupply }}</div>
-      </div>
-      <div>
-        <div class="italic text-right text-xs">{{$t('dashboard.amount')}}:</div>
-        <div>{{ assetAmount }}</div>
-      </div>
-      <div>
-        <div class="italic text-right text-xs">{{$t('dashboard.duration')}}:</div>
-        <div>{{ assetDuration }}</div>
-      </div>
-      <div>
-        <div class="italic text-right text-xs">{{$t('services.divisibility')}}:</div>
-        <div>{{ assetDivisibility }}</div>
-      </div>
-      <div>
-        <div class="italic text-right text-xs">{{$t('services.transferable')}}:</div>
-        <div>
-          <font-awesome-icon v-if="assetTransferable" icon="check-circle" class="h-4 w-4 text-green-600"></font-awesome-icon>
-          <font-awesome-icon v-else icon="times-circle" class="h-4 w-d text-red-600"></font-awesome-icon>
-        </div>
-      </div>
-      <div>
-        <div class="italic text-right text-xs">{{$t('services.supplymutable')}}:</div>
-        <div>
-          <font-awesome-icon v-if="assetMutable" icon="check-circle" class="h-4 w-4 text-green-600"></font-awesome-icon>
-          <font-awesome-icon v-else icon="times-circle" class="h-4 w-d text-red-600"></font-awesome-icon>
-        </div>
-      </div>
-    </div>
-  </div>
-  </div>
-</template>
-<script>
-import { computed, ref, watch } from 'vue';
-// import { useRouter } from "vue-router";
-import PasswordInput from '@/components/PasswordInput.vue';
-import SupplyInput from '@/components/SupplyInput.vue';
-import SelectInputPlugin from '@/components/SelectInputPlugin.vue';
-import { ChainProfileConfig } from "@/models/stores/";
-import { Wallet } from "@/models/wallet";
-import { walletState } from "@/state/walletState";
-import { networkState } from "@/state/networkState";
-import { Currency } from "@/models/currency";
-import { Helper } from '@/util/typeHelper';
-import { ChainUtils } from '@/util/chainUtils';
-import { AssetsUtils } from '@/util/assetsUtils';
-import { WalletUtils } from '@/util/walletUtils';
-import {useI18n} from 'vue-i18n' 
-export default {
-  name: 'ViewServicesAssetsModifySupplyChange_',
-  components: {
-    PasswordInput,
-    SupplyInput,
-    SelectInputPlugin,
-  },
-  setup(){
-    const {t} = useI18n();
-    let maxAmount = 9999999999.999999;
-    const assetOption = ref(null);
-    const showSupplyErr = ref(false);
-    const walletPassword = ref('');
-    const err = ref('');
-    const showMenu = ref(false);
-    const currentSelectedName = ref('');
-    const disabledPassword = ref(false);
-    const disabledSupply = ref(false);
-    const disabledClear = ref(false);
-    const disabledSelectAsset = ref(false);
-    const disabledSelectIncreaseDecrease = ref(false);
-    const walletName = walletState.currentLoggedInWallet.name
-    const passwdPattern = "^[^ ]{8,}$";
-    const showPasswdError = ref(false);
-
-    const cosignerBalanceInsufficient = ref(false);
-    const cosignerAddress = ref('');
-
-    const currencyName = computed(() => networkState.currentNetworkProfile.network.currency.name);
-
-    const lockFund = computed(()=> Helper.convertToExact(networkState.currentNetworkProfileConfig.lockedFundsPerAggregate, networkState.currentNetworkProfile.network.currency.divisibility))
-    const lockFundCurrency = computed(()=> Helper.convertToCurrency(networkState.currentNetworkProfileConfig.lockedFundsPerAggregate, networkState.currentNetworkProfile.network.currency.divisibility))
-
-    const lockFundTxFee = ref(0.0445);
-    const lockFundTotalFee = computed(()=> lockFund.value + lockFundTxFee.value);
-
-
-    const disableCreate = computed(() => !(
-      walletPassword.value.match(passwdPattern) && (supply.value > 0)
-    ));
-
-    const isMultiSig = (address) => {
-      const account = walletState.currentLoggedInWallet.accounts.find((account) => account.address == address);
-      const other = walletState.currentLoggedInWallet.others.find((account) => account.address == address);
-      let isMulti = false;
-      const accountDirectParent = account?account.getDirectParentMultisig():[];
-      const otherDirectParent = other?other.getDirectParentMultisig():[];
-      if((accountDirectParent.length + otherDirectParent.length) > 0){
-        isMulti = true;
-      }
-      return isMulti;
-    };
-
-    const selectedAccName = ref(walletState.currentLoggedInWallet.selectDefaultAccount().name);
-    const selectedAccAdd = ref(walletState.currentLoggedInWallet.selectDefaultAccount().address);
-    const balance = ref(Helper.toCurrencyFormat(walletState.currentLoggedInWallet.selectDefaultAccount().balance, networkState.currentNetworkProfile.network.currency.divisibility));
-    const balanceNumber = ref(maxAmount);
-    const isMultiSigBool = ref(isMultiSig(walletState.currentLoggedInWallet.selectDefaultAccount().address));
-
-    const showNoBalance = ref(false);
-    const showNoAsset = ref(false);
-    const isNotCosigner = computed(() => getMultiSigCosigner.value.list.length == 0 && isMultiSig(selectedAccAdd.value) && !showNoAsset.value);
-
-    const supply = ref('0');
-
-    const accounts = computed( () => {
-      if(walletState.currentLoggedInWallet){
-        if(walletState.currentLoggedInWallet.others){
-          const concatOther = walletState.currentLoggedInWallet.accounts.concat(walletState.currentLoggedInWallet.others)
-          return concatOther;
-        } else{
-          return walletState.currentLoggedInWallet.accounts;
-        }
-      } else{
-        return [];
-      }
-    });
-
-    const moreThanOneAccount = computed(()=>{
-      return accounts.value.length > 1;
-    });
-    const transactionFee = ref('0.000000');
-    const transactionFeeExact = ref(0);
-
-    const ownerPublicAccount = ref(WalletUtils.createPublicAccount(walletState.currentLoggedInWallet.selectDefaultAccount().publicKey, networkState.currentNetworkProfile.network.type));
-
-    const getMultiSigCosigner = computed(() => {
-      return AssetsUtils.getCosignerList(selectedAccAdd.value);
-    });
-
-    const changeSelection = (i) => {
-      assetOption.value.clear();
-      selectedAccName.value = i.name;
-      selectedAccAdd.value = i.address;
-      balance.value = i.balance;
-      showMenu.value = !showMenu.value;
-      currentSelectedName.value = i.name;
-      isMultiSigBool.value = isMultiSig(i.address);
-      ownerPublicAccount.value = WalletUtils.createPublicAccount(i.publicKey, networkState.currentNetworkProfile.network.type);
-      showNoAsset.value = (assetOptions.value.length == 0)?true:false;
-      showNoBalance.value = ((balance.value < totalFee.value) && !isNotCosigner.value && !showNoAsset.value)?true:false;
-    }
-
-    const assetSupply = ref(0);
-    const assetAmount = ref(0);
-    const assetDuration = ref('0 Day');
-    const assetDivisibility = ref(0);
-    const assetTransferable = ref(false);
-    const assetMutable = ref(false);
-    const assetIdString = ref('');
-
-    const changeAsset = (assetId) => {
-      const account = walletState.currentLoggedInWallet.accounts.find((account) => account.address === selectedAccAdd.value)
-      const other = walletState.currentLoggedInWallet.others.find((account) => account.address === selectedAccAdd.value)
-      const selectedAsset = account?account.assets.find((asset) => asset.idHex === assetId):other.assets.find((asset) => asset.idHex === assetId);
-      assetSupply.value =Helper.amountFormatterSimple(selectedAsset.supply, selectedAsset.divisibility);
-      if(selectedAsset.duration){
-        let numDaysleft = Math.ceil(selectedAsset.duration/(24 * 60 * 4));
-        assetDuration.value = numDaysleft + ' Day' + ((numDaysleft>1)?'s':'');
-      }else{
-        assetDuration.value = 'No expiration date';
-      }
-      assetAmount.value = selectedAsset.amount;
-      assetDivisibility.value = selectedAsset.divisibility;
-      assetTransferable.value = selectedAsset.transferable;
-      assetMutable.value = selectedAsset.supplyMutable;
-      transactionFee.value = Helper.amountFormatterSimple(AssetsUtils.getMosaicSupplyChangeTransactionFee(networkState.currentNetworkProfile.network.type, networkState.currentNetworkProfile.generationHash, assetIdString, selectIncreaseDecrease.value, supply.value, assetDivisibility.value), networkState.currentNetworkProfile.network.currency.divisibility);
-      transactionFeeExact.value = Helper.convertToExact(AssetsUtils.getMosaicSupplyChangeTransactionFee(networkState.currentNetworkProfile.network.type, networkState.currentNetworkProfile.generationHash, assetIdString, selectIncreaseDecrease.value, supply.value, assetDivisibility.value), networkState.currentNetworkProfile.network.currency.divisibility);
-      balanceNumber.value = selectIncreaseDecrease.value=='increase'?maxAmount:parseFloat(assetSupply.value);
-      showSupplyErr.value = supply.value>balanceNumber.value;
-    };
-
-    const assetOptions = computed(() => {
-      return AssetsUtils.getOwnedAssetsPermutable(selectedAccAdd.value);
-    });
-
-    const increaseDecreaseOption = () => {
-      let action = [];
-      action.push({value: 'decrease', label: t('services.decrease')});
-      action.push({value: 'increase', label: t('services.increase')});
-      return action;
-    };
-
-    const selectAsset = ref('');
-    const selectIncreaseDecrease = ref('increase');
-
-    const clearInput = () => {
-      walletPassword.value = '';
-      supply.value = '0';
-      assetOption.value.clear();
-    };
-
-    const modifyMosaic = () => {
-       if(cosigner.value){
-         AssetsUtils.changeAssetSupplyMultiSig(cosigner.value, walletPassword.value, networkState.currentNetworkProfile.network.type, networkState.currentNetworkProfile.generationHash, selectAsset.value, selectIncreaseDecrease.value, supply.value, assetDivisibility.value, selectedAccAdd.value);
-       }else{
-        AssetsUtils.changeAssetSupply(selectedAccAdd.value, walletPassword.value, networkState.currentNetworkProfile.network.type, networkState.currentNetworkProfile.generationHash, selectAsset.value, selectIncreaseDecrease.value, supply.value, assetDivisibility.value);
-       }
-      clearInput();
-    };
-
-    watch(selectIncreaseDecrease, (n) => {
-      if(selectAsset.value){
-        transactionFee.value = Helper.amountFormatterSimple(AssetsUtils.getMosaicSupplyChangeTransactionFee(networkState.currentNetworkProfile.network.type, networkState.currentNetworkProfile.generationHash, selectAsset.value, n, supply.value, assetDivisibility.value), networkState.currentNetworkProfile.network.currency.divisibility);
-        transactionFeeExact.value = Helper.convertToExact(AssetsUtils.getMosaicSupplyChangeTransactionFee(networkState.currentNetworkProfile.network.type, networkState.currentNetworkProfile.generationHash, selectAsset.value, n, supply.value, assetDivisibility.value), networkState.currentNetworkProfile.network.currency.divisibility);
-        balanceNumber.value = (n=='increase'?maxAmount:parseFloat(assetSupply.value));
-      }else{
-        balanceNumber.value = (n=='increase'?maxAmount:0);
-      }
-      showSupplyErr.value = supply.value>balanceNumber.value;
-    });
-
-    watch(supply, (n) => {
-      if(selectAsset.value){
-        transactionFee.value = Helper.amountFormatterSimple(AssetsUtils.getMosaicSupplyChangeTransactionFee(networkState.currentNetworkProfile.network.type, networkState.currentNetworkProfile.generationHash, selectAsset.value, selectIncreaseDecrease.value, n, assetDivisibility.value), networkState.currentNetworkProfile.network.currency.divisibility);
-        transactionFeeExact.value = Helper.convertToExact(AssetsUtils.getMosaicSupplyChangeTransactionFee(networkState.currentNetworkProfile.network.type, networkState.currentNetworkProfile.generationHash, selectAsset.value, selectIncreaseDecrease.value, n, assetDivisibility.value), networkState.currentNetworkProfile.network.currency.divisibility);
-      }
-    });
-
-    const totalFee = computed(() => {
-      // if multisig
-      if(isMultiSig(selectedAccAdd.value)){
-        return parseFloat(lockFundTotalFee.value) + transactionFeeExact.value;
-      }else{
-        return transactionFeeExact.value;
-      }
-    });
-
-    const setFormInput = (isValidate) => {
-      disabledPassword.value = isValidate;
-      disabledSupply.value = isValidate;
-      disabledSelectAsset.value = isValidate;
-      disabledSelectIncreaseDecrease.value = isValidate;
-    };
-
-    const clearAsset = () => {
-      assetSupply.value = 0;
-      assetAmount.value = 0;
-      assetDuration.value = '0 Day';
-      assetDivisibility.value = 0;
-      assetTransferable.value = false;
-      assetMutable.value = false;
-    };
-
-    watch(totalFee, (n) => {
-      if(balance.value < n){
-        if(!showNoAsset.value){
-          if(!isNotCosigner.value){
-            showNoBalance.value = true;
-          }
-        }
-        setFormInput(true);
-      }else{
-        showNoBalance.value = false;
-        setFormInput(false);
-      }
-    });
-
-    watch(showNoAsset, (n) => {
-      if(n){
-        setFormInput(true);
-      }else{
-        setFormInput(false);
-      }
-    });
-
-    watch(showNoBalance, (n) => {
-      if(n){
-        setFormInput(true);
-      }else{
-        setFormInput(false);
-      }
-    });
-
-    watch(isNotCosigner, (n) => {
-      if(n){
-        setFormInput(true)
-      }else{
-        setFormInput(false);
-      }
-    });
-
-    const cosigner = ref('');
-    // get cosigner
-    watch(getMultiSigCosigner, (n) => {
-      // if it is a multisig
-      if(n.list.length > 0){
-        if(n.list.length > 1){
-          cosigner.value = cosignerAddress.value;
-        }else{
-          cosigner.value = n.list[0].address;
-        }
-      }else{
-        cosigner.value = '';
-      }
-    });
-
-    return {
-      assetOption,
-      accounts,
-      moreThanOneAccount,
-      showMenu,
-      currentSelectedName,
-      selectedAccName,
-      selectedAccAdd,
-      balance,
-      balanceNumber,
-      showNoBalance,
-      lockFundTotalFee,
-      showNoAsset,
-      showSupplyErr,
-      lockFundCurrency,
-      lockFundTxFee,
-      err,
-      walletPassword,
-      disableCreate,
-      clearInput,
-      showPasswdError,
-      changeSelection,
-      supply,
-      disabledPassword,
-      disabledSupply,
-      disabledClear,
-      currencyName,
-      isMultiSig,
-      isMultiSigBool,
-      assetOptions,
-      increaseDecreaseOption,
-      selectAsset,
-      selectIncreaseDecrease,
-      modifyMosaic,
-      transactionFee,
-      transactionFeeExact,
-      changeAsset,
-      assetSupply,
-      assetAmount,
-      assetDuration,
-      assetDivisibility,
-      assetTransferable,
-      assetMutable,
-      getMultiSigCosigner,
-      cosignerBalanceInsufficient,
-      cosignerAddress,
-      isNotCosigner,
-      disabledSelectAsset,
-      disabledSelectIncreaseDecrease,
-      walletName,
-      clearAsset,
-    }
-  },
-
-}
-</script>
-<style scoped lang="scss">
-
-.slide-enter-active {
-   -moz-transition-duration: 1s;
-   -webkit-transition-duration: 1s;
-   -o-transition-duration: 1s;
-   transition-duration: 1s;
-   -moz-transition-timing-function: ease-in;
-   -webkit-transition-timing-function: ease-in;
-   -o-transition-timing-function: ease-in;
-   transition-timing-function: ease-in;
-}
-
-.slide-leave-active {
-   -moz-transition-duration: 1s;
-   -webkit-transition-duration: 1s;
-   -o-transition-duration: 1s;
-   transition-duration: 1s;
-   -moz-transition-timing-function: cubic-bezier(0, 1, 0.5, 1);
-   -webkit-transition-timing-function: cubic-bezier(0, 1, 0.5, 1);
-   -o-transition-timing-function: cubic-bezier(0, 1, 0.5, 1);
-   transition-timing-function: cubic-bezier(0, 1, 0.5, 1);
-}
-
-.slide-enter-to, .slide-leave-from {
-  max-height: 1000px;
-  overflow: hidden;
-}
-
-.slide-enter-from, .slide-leave-to {
-   overflow: hidden;
-   max-height: 0;
-}
-
-.optionDiv:hover{
-  background: #D9EBFF;
-}
-
-.property-table{
-  > div {
-    margin-bottom: 7px;
-    > div{
-      display: inline-block;
-    }
-    > div:first-child{
-      width: 75px;
-      margin-right: 15px;
-    }
-  }
-}
-
-@media(min-width: 1024px){
-  .property-table{
-    > div {
-      > div:first-child{
-        width: 90px;
-      }
-    }
-  }
-}
-</style>
